+++ src/server/RLCreatorLayer.cpp
@@ -346,7 +346,7 @@ void RLCreatorLayer::onLeaderboard(CCObject* sender) {
 void RLCreatorLayer::onFeaturedLayouts(CCObject* sender) {
       web::WebRequest()
           .param("type", 2)
-          .param("amount", 1000)
+          .param("amount", 100)
           .get("https://gdrate.arcticwoof.xyz/getLevels")
           .listen([this](web::WebResponse* res) {
                 if (res && res->ok()) {
@@ -401,7 +401,7 @@ void RLCreatorLayer::onFeaturedLayouts(CCObject* sender) {
 void RLCreatorLayer::onNewRated(CCObject* sender) {
       web::WebRequest()
           .param("type", 3)
-          .param("amount", 1000)
+          .param("amount", 100)
           .get("https://gdrate.arcticwoof.xyz/getLevels")
           .listen([this](web::WebResponse* res) {
                 if (res && res->ok()) {
@@ -456,7 +456,7 @@ void RLCreatorLayer::onNewRated(CCObject* sender) {
 void RLCreatorLayer::onSendLayouts(CCObject* sender) {
       web::WebRequest()
           .param("type", 1)
-          .param("amount", 1000)
+          .param("amount", 100)
           .get("https://gdrate.arcticwoof.xyz/getLevels")
           .listen([this](web::WebResponse* res) {
                 if (res && res->ok()) {